+++ motrack/tracker/matching/factory.py
@@ -5,6 +5,16 @@ Use `ASSOCIATION_CATALOG.register` to extend supported tracker association algor
 from typing import Dict, Any
 
 from motrack.tracker.matching.algorithms.base import AssociationAlgorithm
+# noinspection PyUnresolvedReferences
+from motrack.tracker.matching.algorithms.biou import HungarianCBIoU, HungarianBIoU
+# noinspection PyUnresolvedReferences
+from motrack.tracker.matching.algorithms.dcm import DCMIoU, MoveDCM
+# noinspection PyUnresolvedReferences
+from motrack.tracker.matching.algorithms.iou import IoUAssociation
+# noinspection PyUnresolvedReferences
+from motrack.tracker.matching.algorithms.move import Move
+# noinspection PyUnresolvedReferences
+from motrack.tracker.matching.algorithms.reid_iou import ReidIoUAssociation
 from motrack.tracker.matching.catalog import ASSOCIATION_CATALOG