+++ motrack/tracker/trackers/algorithms/byte.py
@@ -9,7 +9,7 @@ import numpy as np
 
 from motrack.library.cv.bbox import PredBBox
 from motrack.tracker.matching import association_factory
-from motrack.tracker.trackers.algorithms.motion import MotionBasedTracker
+from motrack.tracker.trackers.algorithms.motion_reid import MotionReIDBasedTracker
 from motrack.tracker.trackers.catalog import TRACKER_CATALOG
 from motrack.tracker.trackers.utils import remove_duplicates
 from motrack.tracker.tracklet import Tracklet, TrackletState
@@ -17,7 +17,7 @@ from motrack.utils.collections import unpack_n
 
 
 @TRACKER_CATALOG.register('byte')
-class ByteTracker(MotionBasedTracker):
+class ByteTracker(MotionReIDBasedTracker):
     """
     ByteTrack algorithm.
 
@@ -40,6 +40,8 @@ class ByteTracker(MotionBasedTracker):
         filter_params: Optional[dict] = None,
         cmc_name: Optional[str] = None,
         cmc_params: Optional[dict] = None,
+        reid_name: Optional[str] = None,
+        reid_params: Optional[str] = None,
         high_matcher_algorithm: str = 'default',
         high_matcher_params: Optional[Dict[str, Any]] = None,
         low_matcher_algorithm: str = 'default',
@@ -56,11 +58,17 @@ class ByteTracker(MotionBasedTracker):
         if filter_params is None:
             filter_params = {}
 
+        new_tracklet_detection_threshold = new_tracklet_detection_threshold if new_tracklet_detection_threshold is not None \
+            else detection_threshold
         super().__init__(
             filter_name=filter_name,
             filter_params=filter_params,
             cmc_name=cmc_name,
-            cmc_params=cmc_params
+            cmc_params=cmc_params,
+            reid_name=reid_name,
+            reid_params=reid_params,
+
+            new_tracklet_detection_threshold=new_tracklet_detection_threshold
         )
 
         if high_matcher_algorithm == 'default':
@@ -97,8 +105,6 @@ class ByteTracker(MotionBasedTracker):
         self._remember_threshold = remember_threshold
         self._use_observation_if_lost = use_observation_if_lost
         self._detection_threshold = detection_threshold
-        self._new_tracklet_detection_threshold = new_tracklet_detection_threshold \
-            if new_tracklet_detection_threshold is not None else detection_threshold
         self._duplicate_iou_threshold = duplicate_iou_threshold
 
         # State
@@ -111,6 +117,7 @@ class ByteTracker(MotionBasedTracker):
         prior_tracklet_bboxes: List[PredBBox],
         detections: List[PredBBox],
         frame_index: int,
+        objects_features: Optional[np.ndarray] = None,
         inplace: bool = True,
         frame: Optional[np.ndarray] = None
     ) -> List[Tracklet]:
@@ -123,8 +130,10 @@ class ByteTracker(MotionBasedTracker):
         # (2) Match high detections with tracklets with states ACTIVE and LOST using HighMatchAlgorithm
         tracklets_active_and_lost_indices, tracklets_active_and_lost, tracklets_active_and_lost_bboxes = \
             unpack_n([(i, t, t_bbox) for i, (t, t_bbox) in enumerate(zip(tracklets, prior_tracklet_bboxes)) if t.is_tracked], n=3)
+        reid_features_active_and_lost = objects_features[tracklets_active_and_lost_indices] if objects_features is not None else None
         high_matches, remaining_tracklet_indices, high_unmatched_detections_indices = \
-            self._high_match(tracklets_active_and_lost_bboxes, high_detections, tracklets_active_and_lost)
+            self._high_match(tracklets_active_and_lost_bboxes, high_detections,
+                             object_features=reid_features_active_and_lost, tracklets=tracklets_active_and_lost)
         high_matches = [(tracklets_active_and_lost_indices[t_i], high_det_indices[d_i]) for t_i, d_i in high_matches]
         high_unmatched_detections_indices = [high_det_indices[d_i] for d_i in high_unmatched_detections_indices]
         remaining_tracklet_bboxes = [tracklets_active_and_lost_bboxes[t_i] for t_i in remaining_tracklet_indices]
@@ -135,8 +144,10 @@ class ByteTracker(MotionBasedTracker):
         remaining_active_tracklet_indices, remaining_active_tracklets, remaining_active_tracklet_bboxes = \
             unpack_n([(i, t, t_bbox) for i, t, t_bbox in zip(remaining_tracklet_indices, remaining_tracklets, remaining_tracklet_bboxes)
                   if t.state == TrackletState.ACTIVE], n=3)
+        reid_features_remaining_active = objects_features[remaining_active_tracklet_indices] if objects_features is not None else None
         low_matches, low_unmatched_tracklet_indices, _ = \
-            self._low_match(remaining_active_tracklet_bboxes, low_detections, remaining_active_tracklets)
+            self._low_match(remaining_active_tracklet_bboxes, low_detections,
+                            object_features=reid_features_remaining_active, tracklets=remaining_active_tracklets)
         low_matches = [(remaining_active_tracklet_indices[t_i], low_det_indices[d_i]) for t_i, d_i in low_matches]
         unmatched_tracklet_indices = [remaining_active_tracklet_indices[t_i] for t_i in low_unmatched_tracklet_indices]
 
@@ -150,28 +161,21 @@ class ByteTracker(MotionBasedTracker):
         remaining_high_detection_indices = high_unmatched_detections_indices
         tracklets_new_indices, tracklets_new, tracklets_new_bboxes = \
             unpack_n([(i, t, t_bbox) for i, (t, t_bbox) in enumerate(zip(tracklets, prior_tracklet_bboxes)) if t.state == TrackletState.NEW], n=3)
+        reid_features_new = objects_features[tracklets_new_indices] if objects_features is not None else None
         new_matches, new_unmatched_tracklets_indices, new_unmatched_detections_indices = \
-            self._new_match(tracklets_new_bboxes, remaining_high_detections, tracklets_new)
+            self._new_match(tracklets_new_bboxes, remaining_high_detections,
+                            object_features=reid_features_new, tracklets=tracklets_new)
         new_matches = [(tracklets_new_indices[t_i], high_unmatched_detections_indices[d_i]) for t_i, d_i in new_matches]
         new_unmatched_tracklets_indices = [tracklets_new_indices[t_i] for t_i in new_unmatched_tracklets_indices]
         new_unmatched_detections_indices = [remaining_high_detection_indices[d_i] for d_i in new_unmatched_detections_indices]
 
         # (6) Initialize new tracklets from unmatched high detections
-        new_tracklets: List[Tracklet] = []
-        for d_i in new_unmatched_detections_indices:
-            detection = detections[d_i]
-            if self._new_tracklet_detection_threshold is not None and detection.conf < self._new_tracklet_detection_threshold:
-                continue
-
-            new_tracklet = Tracklet(
-                bbox=detection if inplace else copy.deepcopy(detection),
-                frame_index=frame_index,
-                _id=self._next_id,
-                state=TrackletState.NEW if frame_index > 1 else TrackletState.ACTIVE
-            )
-            self._next_id += 1
-            new_tracklets.append(new_tracklet)
-            self._initiate(new_tracklet.id, detection)
+        new_tracklets = self._create_new_tracklets(
+            detections=[detections[d_i] for d_i in new_unmatched_detections_indices],
+            frame_index=frame_index,
+            objects_features=objects_features[new_unmatched_detections_indices] if objects_features is not None else None,
+            inplace=inplace
+        )
 
         all_matches = high_matches + low_matches + new_matches