+++ motrack/tracker/trackers/algorithms/motion_reid.py
@@ -5,15 +5,17 @@ from typing import Tuple, Optional, List
 
 import numpy as np
 from abc import ABC, abstractmethod
+import copy
 
 from motrack.filter import filter_factory
 from motrack.library.cv.bbox import PredBBox, BBox
 from motrack.tracker.trackers.algorithms.base import Tracker
-from motrack.tracker.tracklet import Tracklet, TrackletState
+from motrack.tracker.tracklet import Tracklet, TrackletState, TrackletCommonData
 from motrack.cmc import cmc_factory
+from motrack.reid import reid_inference_factory
 
 
-class MotionBasedTracker(Tracker, ABC):
+class MotionReIDBasedTracker(Tracker, ABC):
     """
     Baseline Sort tracker components:
     - ReID: HungarianIOU
@@ -25,7 +27,10 @@ class MotionBasedTracker(Tracker, ABC):
         filter_name: str,
         filter_params: dict,
         cmc_name: Optional[str] = None,
-        cmc_params: Optional[dict] = None
+        cmc_params: Optional[dict] = None,
+        reid_name: Optional[str] = None,
+        reid_params: Optional[str] = None,
+        new_tracklet_detection_threshold: Optional[float] = None,
     ):
         """
         Args:
@@ -33,19 +38,29 @@ class MotionBasedTracker(Tracker, ABC):
             filter_params: Filter params
             cmc_name: CMC name
             cmc_params: CMC params
+            reid_name: ReID name
+            reid_params: ReID params
         """
         super().__init__()
 
         self._filter = filter_factory(filter_name, filter_params)
 
+        self._cmc = None
         if cmc_name is not None:
             cmc_params = {} if cmc_params is None else cmc_params
             self._cmc = cmc_factory(cmc_name, cmc_params)
-        else:
-            self._cmc = None
+
+        self._reid = None
+        if reid_name is not None:
+            reid_params = {} if reid_params is None else reid_params
+            self._reid = reid_inference_factory(reid_name, reid_params)
+
+        # Hyperparameters
+        self._new_tracklet_detection_threshold = new_tracklet_detection_threshold
 
         # State
         self._filter_states = {}
+        self._next_id = 0
 
     @staticmethod
     def _raw_to_bbox(tracklet: Tracklet, raw: np.ndarray, conf: Optional[float] = None) -> PredBBox:
@@ -142,6 +157,34 @@ class MotionBasedTracker(Tracker, ABC):
         """
         self._filter_states.pop(tracklet_id)
 
+    def _create_new_tracklets(
+        self,
+        detections: List[PredBBox],
+        frame_index: int,
+        objects_features: Optional[np.ndarray] = None,
+        inplace: bool = True
+    ) -> List[Tracklet]:
+        new_tracklets: List[Tracklet] = []
+        for d_i, detection in enumerate(detections):
+            if self._new_tracklet_detection_threshold is not None and detection.conf < self._new_tracklet_detection_threshold:
+                continue
+
+            new_tracklet = Tracklet(
+                bbox=detection if inplace else copy.deepcopy(detection),
+                frame_index=frame_index,
+                _id=self._next_id,
+                state=TrackletState.NEW if frame_index > 1 else TrackletState.ACTIVE
+            )
+            self._next_id += 1
+            new_tracklets.append(new_tracklet)
+            self._initiate(new_tracklet.id, detection)
+
+            # Set initial tracklet feature if ReId model is used
+            if objects_features is not None:
+                new_tracklet.set(TrackletCommonData.APPEARANCE, objects_features[d_i])
+
+        return new_tracklets
+
     def _perform_cmc(self, frame: np.ndarray, frame_index: int, bboxes: List[PredBBox]) -> List[PredBBox]:
         if self._cmc is None:
             return bboxes
@@ -160,6 +203,13 @@ class MotionBasedTracker(Tracker, ABC):
 
         return corrected_bboxes
 
+
+    def _extract_reid_features(self, frame: np.ndarray, frame_index: int, bboxes: List[PredBBox]) -> Optional[np.ndarray]:
+        if self._reid is None:
+            return None
+
+        return self._reid.extract_objects_features(frame, bboxes, frame_index=frame_index, scene=self._scene)
+
     def track(
         self,
         tracklets: List[Tracklet],
@@ -168,8 +218,11 @@ class MotionBasedTracker(Tracker, ABC):
         inplace: bool = True,
         frame: Optional[np.ndarray] = None
     ) -> List[Tracklet]:
-        tracklets, prior_tracklet_bboxes = self._track_predict(tracklets, frame_index - 1, frame=frame)
-        return self._track(tracklets, prior_tracklet_bboxes, detections, frame_index - 1, inplace=inplace, frame=frame)
+        frame_index -= 1
+        tracklets, prior_tracklet_bboxes = self._track_predict(tracklets, frame_index, frame=frame)
+        objects_features = self._extract_reid_features(frame, frame_index, detections)
+        return self._track(tracklets, prior_tracklet_bboxes, detections, frame_index,
+                           objects_features=objects_features, inplace=inplace, frame=frame)
 
     def _track_predict(self, tracklets: List[Tracklet], frame_index: int, frame: Optional[np.ndarray] = None):
         """
@@ -203,6 +256,7 @@ class MotionBasedTracker(Tracker, ABC):
         prior_tracklet_bboxes: List[PredBBox],
         detections: List[PredBBox],
         frame_index: int,
+        objects_features: Optional[np.ndarray] = None,
         inplace: bool = True,
         frame: Optional[np.ndarray] = None
     ) -> List[Tracklet]:
@@ -214,6 +268,7 @@ class MotionBasedTracker(Tracker, ABC):
             prior_tracklet_bboxes: Tracklet prior position bboxes
             detections: Lists of new detections
             frame_index: Current frame number
+            objects_features: Object appearance features
             inplace: Perform inplace transformations on tracklets and bboxes
             frame: Pass frame in case CMC or appearance based association is used